Fixup: Unit Tests and input file
The integration method on python was changed to reduce the time the tests take.
The tolerance on the python tests had to be raised to 1e-2.
(the integration term some times fails at 1e-3)
The domain on the input file was expanded to see the waves.
(initialization time is the same)